+++ src/netspresso_trainer/losses/common.py
@@ -8,7 +8,7 @@ from torch import Tensor
 
 class CrossEntropyLoss(nn.Module):
     def __init__(self, weight: Optional[Tensor]=None, size_average=None, ignore_index: int=-100,
-                 reduce=None, label_smoothing: float=0.0):
+                 reduce=None, label_smoothing: float=0.0, **kwargs):
         super(CrossEntropyLoss, self).__init__()
         self.loss_fn = nn.CrossEntropyLoss(weight=weight, size_average=size_average, ignore_index=ignore_index,
                                            reduce=reduce, reduction='mean', label_smoothing=label_smoothing)
@@ -21,7 +21,7 @@ class CrossEntropyLoss(nn.Module):
 
 
 class SigmoidFocalLoss(nn.Module):
-    def __init__(self, alpha, gamma):
+    def __init__(self, alpha, gamma, **kwargs):
         super(SigmoidFocalLoss, self).__init__()
         self.alpha = alpha
         self.gamma = gamma