+++ src/netspresso_trainer/pipelines/base.py
@@ -33,7 +33,7 @@ class BasePipeline(ABC):
 
     @property
     def sample_input(self):
-        return torch.randn((1, 3, self.conf.augmentation.img_size, self.conf.augmentation.img_size))
+        return torch.randn((1, 3, self.conf.logging.onnx_input_size[0], self.conf.logging.onnx_input_size[1]))
 
     def log_results(
         self,