+++ src/netspresso_trainer/trainer_common.py
@@ -69,7 +69,6 @@ def train_common(
             conf.model, task, train_dataset.num_classes,
             model_checkpoint=conf.model.checkpoint.path,
             use_pretrained=conf.model.checkpoint.use_pretrained,
-            img_size=conf.augmentation.img_size
         )
 
     model = model.to(device=devices)